server/requirements.txt
python-dotenv
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
alembic
httpx
langchain
langgraph
langchain_core
langchain_community
langchain-google-genai
langserve
requests
sse_starlette
ipykernel
google-genai
soundfile
pyaudio
google-adk
websockets
ipython
elevenlabs
streamlit
google-cloud-aiplatform
langchain-groq
langchain-google-vertexai
PyMuPDF
fuzzywuzzy
python-docx
typing
sentence-transformers 
pypdf
ftfy 
regex 
tqdm
gradio
selenium
kagglehub
folium
matplotlib
plotly
pandas 
geopandas 
duckduckgo-search
langchain-huggingface
tiktoken
numpy
beautifulsoup4
streamlit-agraph 
graphviz
pdfplumber 
pypdfium2 
unstructured
unstructured[docx]
textloader
google-search-results
redis
chromadb
langchain-openai
langchain-anthropic
langchain-groq
langchain-mistralai
langchain-mongodb
langchain-docling
langchain-tavily
tiktoken
lxml
unstructured[docx,pdf,xlsx]
python-docx
pypdf
langchain-chroma
